refactor(subnets): refactored Fabrics structure
