+++ src/app/subnets/views/Fabrics/components/DeleteFabric/DeleteFabric.tsx
