+++ src/app/subnets/views/Fabrics/components/DeleteFabric/index.ts
