--- src/app/subnets/views/Fabrics/components/FabricDetailsHeader/constants.ts
@@ -1,11 +1,0 @@
-import type { ValueOf } from "@canonical/react-components";
-
-import type { SidePanelContent } from "@/app/base/types";
-
-export const FabricDetailsSidePanelViews = {
-  DELETE_FABRIC: ["", "deleteFabric"],
-} as const;
-
-export type FabricDetailsSidePanelContent = SidePanelContent<
-  ValueOf<typeof FabricDetailsSidePanelViews>
->;
